opcua::client

Module prelude

Re-exports

pub use crate::types::status_code::StatusCode;

pub use crate::core::prelude::*;

pub use crate::crypto::*;

pub use crate::types::service_types::*;

Structs

Client

The Client defines a connection that can be used to to get end points or establish one or more sessions with an OPC UA server. It is constructed via a ClientBuilder or from a described configuration ClientConfig that could be deserialized from file.

ClientBuilder

The ClientBuilder is a builder for producing a Client. It is an alternative to constructing a ClientConfig from file or from scratch.

ClientConfig

Client OPC UA configuration

ClientEndpoint

Describes an endpoint, it’s url security policy, mode and user token

ClientUserToken

ConnectionStatusCallback

This is a concrete implementation of OnConnectionStatusChange that calls the supplied function.

DataChangeCallback

This is a concrete implementation of OnSubscriptionNotification that calls a function when a data change occurs.

EventCallback

This is a concrete implementation of OnSubscriptionNotification that calls a function when an event occurs.

MonitoredItem

Performance

Session

A session of the client. The session is associated with an endpoint and maintains a state when it is active. The Session struct provides functions for all the supported request types in the API.

SessionClosedCallback

This is a concrete implementation of OnSessionClosed that will call the supplied function.

SessionInfo

Information about the server endpoint, security policy, security mode and user identity that the session will will use to establish a connection.

Enums

HistoryReadAction

Enumeration used with Session::history_read()

HistoryUpdateAction

Enumeration used with Session::history_update()

IdentityToken

SessionCommand

A Session runs in a loop, which can be terminated by sending it a SessionCommand.

Constants

ANONYMOUS_USER_TOKEN_ID

Traits

AttributeService

Attribute Service set

DiscoveryService

Discovery Service set

MethodService

Method Service set

MonitoredItemService

NodeManagementService

NodeManagement Service set

OnConnectionStatusChange

The OnConnectionStatusChange trait can be used to register on the session to be notified of connection status change notifications.

OnSessionClosed

The OnSessionClosed trait can be used to register on a session and called to notify the client that the session has closed.

OnSubscriptionNotification

The OnSubscriptionNotification trait is the callback registered along with a new subscription to receive subscription notification callbacks.

SecureChannelService

SecureChannel Service set

Service

SessionService

Session Service set

SubscriptionService

ViewService

View Service set