opcua::client::prelude

Struct ClientConfig

pub struct ClientConfig {

    pub application_name: String,
    pub application_uri: String,
    pub product_uri: String,
    pub create_sample_keypair: bool,
    pub certificate_path: Option<PathBuf>,
    pub private_key_path: Option<PathBuf>,
    pub trust_server_certs: bool,
    pub verify_server_certs: bool,
    pub pki_dir: PathBuf,
    pub preferred_locales: Vec<String>,
    pub default_endpoint: String,
    pub user_tokens: BTreeMap<String, ClientUserToken>,
    pub endpoints: BTreeMap<String, ClientEndpoint>,
    pub decoding_options: DecodingOptions,
    pub session_retry_limit: i32,
    pub session_retry_interval: u32,
    pub session_timeout: u32,
    pub performance: Performance,
    pub session_name: String,

}

Client OPC UA configuration

Fields
application_name: String

Name of the application that the client presents itself as to the server

application_uri: String

The application uri

product_uri: String

Product uri

create_sample_keypair: bool

Autocreates public / private keypair if they donโt exist. For testing/samples only since you do not have control of the values

certificate_path: Option<PathBuf>

Custom certificate path, to be used instead of the default .der certificate path

private_key_path: Option<PathBuf>

Custom private key path, to be used instead of the default private key path

trust_server_certs: bool

Auto trusts server certificates. For testing/samples only unless youโre sure what youโre doing.

verify_server_certs: bool

Verify server certificates. For testing/samples only unless youโre sure what youโre doing.

pki_dir: PathBuf

PKI folder, either absolute or relative to executable

preferred_locales: Vec<String>

Preferred locales

default_endpoint: String

Identifier of the default endpoint

user_tokens: BTreeMap<String, ClientUserToken>

User tokens

endpoints: BTreeMap<String, ClientEndpoint>

List of end points

decoding_options: DecodingOptions

Decoding options used for serialization / deserialization

session_retry_limit: i32

Max retry limit -1, 0 or number

session_retry_interval: u32

Retry interval in milliseconds

session_timeout: u32

Session timeout period in milliseconds

performance: Performance

Client performance settings

session_name: String

Session name

Implementations

impl ClientConfig

pub const PKI_DIR: &'static str = "pki"

The default PKI directory

pub fn new<T>(application_name: T, application_uri: T) -> Self
where T: Into<String>,