opcua::client::prelude

Struct Session

pub struct Session { /* private fields */ }

A session of the client. The session is associated with an endpoint and maintains a state when it is active. The Session struct provides functions for all the supported request types in the API.

Note that not all servers may support all service requests and calling an unsupported API may cause the connection to be dropped. Your client is expected to know the capabilities of the server it is calling to avoid this.

Implementations

impl Session

pub fn connect_and_activate(&mut self) -> Result<(), StatusCode>

Connects to the server, creates and activates a session. If there is a failure, it will be communicated by the status code in the result.

Returns

Ok(()) - connection has happened and the session is activated

Err(StatusCode) - reason for failure

pub fn set_session_retry_policy( &mut self, session_retry_policy: SessionRetryPolicy, )

Sets the session retry policy that dictates what this session will do if the connection fails or goes down. The retry policy enables the session to retry a connection on an interval up to a maxmimum number of times.

Arguments

session_retry_policy - the session retry policy to use

pub fn set_session_closed_callback<CB>(&mut self, session_closed_callback: CB)
where CB: OnSessionClosed + Send + Sync + 'static,

Register a callback to be notified when the session has been closed.

Arguments

session_closed_callback - the session closed callback

pub fn set_connection_status_callback<CB>( &mut self, connection_status_callback: CB, )
where CB: OnConnectionStatusChange + Send + Sync + 'static,

Registers a callback to be notified when the session connection status has changed. This will be called if connection status changes from connected to disconnected or vice versa.

Arguments

connection_status_callback - the connection status callback.

pub fn reconnect_and_activate(&mut self) -> Result<(), StatusCode>

Reconnects to the server and tries to activate the existing session. If there is a failure, it will be communicated by the status code in the result. You should not call this if there is a session retry policy associated with the session.

Reconnecting will attempt to transfer or recreate subscriptions that were on the old session before it terminated.

Returns

Ok(()) - reconnection has happened and the session is activated

Err(StatusCode) - reason for failure

pub fn connect(&self) -> Result<(), StatusCode>

Connects to the server using the retry policy to repeat connecting until such time as it succeeds or the policy says to give up. If there is a failure, it will be communicated by the status code in the result.

pub fn connect_no_retry(&self) -> Result<(), StatusCode>

Connects to the server using the configured session arguments. No attempt is made to retry the connection if the attempt fails. If there is a failure, it will be communicated by the status code in the result.

Returns

Ok(()) - connection has happened

Err(StatusCode) - reason for failure

pub fn disconnect(&self)

Disconnect from the server. Disconnect is an explicit command to drop the socket and throw away all state information. If you disconnect you cannot reconnect to your existing session or retrieve any existing subscriptions.

pub fn is_connected(&self) -> bool

Test if the session is in a connected state

Returns

true - Session is connected

false - Session is not connected

pub fn run(session: Arc<RwLock<Session>>)

Synchronously runs a polling loop over the supplied session. Running a session performs periodic actions such as receiving messages, processing subscriptions, and recovering from connection errors. The run function will return if the session is disconnected and cannot be reestablished.

Arguments

session - the session to run ynchronously

pub fn run_async(session: Arc<RwLock<Session>>) -> Sender<SessionCommand>

Runs the session asynchronously on a new thread. The function returns immediately and gives a caller a Sender that can be used to send a message to the session to cause it to terminate. Do not drop this sender (i.e. make sure to bind it to a variable with sufficient lifetime) or the session will terminate as soon as you do.

Running a session performs periodic actions such as receiving messages, processing subscriptions, and recovering from. connection errors. The session will terminate by itself if it is disconnected and cannot be reestablished. It will terminate if the sender is dropped or if sent a ClientCommand to terminate. caller to this function can monitor the status of the session through state calls to know when this happens.

Arguments

session - the session to run asynchronously

Returns

oneshot::Sender<ClientCommand> - A sender that allows the caller to send a message to the run loop to cause it to stop. Note that dropping the sender, i.e. not binding it to a variable will also cause the loop to stop.

pub async fn session_task( session: Arc<RwLock<Session>>, sleep_interval: u64, rx: Receiver<SessionCommand>, )

The asynchronous main session loop. This is the function that processes responses and keeps the session alive. Note that while the client normally calls run() or run_loop() to invoke this, there may be situations where the client wishes to directly use this function, for example if the client has its own Tokio runtime and prefers to spawn the task with that.

pub fn run_loop( session: Arc<RwLock<Session>>, sleep_interval: u64, rx: Receiver<SessionCommand>, )

The main running loop for a session. This is used by run() and run_async() to run continuously until a signal is received to terminate.

Arguments

session - The session

sleep_interval - An internal polling timer in ms

rx - A receiver that the task uses to receive a quit command directly from the caller.

pub async fn poll(&mut self) -> Result<bool, ()>

Polls on the session which basically dispatches any pending async responses, attempts to reconnect if the client is disconnected from the client and sleeps a little bit if nothing needed to be done.

Arguments

sleep_for - the period of time in milliseconds that poll should sleep for if it performed no action.

Returns

true - if an action was performed during the poll

false - if no action was performed during the poll and the poll slept

pub fn delete_all_subscriptions( &self, ) -> Result<Vec<(u32, StatusCode)>, StatusCode>

Deletes all subscriptions by sending a DeleteSubscriptionsRequest to the server with ids for all subscriptions.

Returns

Ok(Vec<(u32, StatusCode)>) - List of (id, status code) result for delete action on each id, Good or BadSubscriptionIdInvalid

Err(StatusCode) - Status code reason for failure

pub fn close_session_and_delete_subscriptions(&self) -> Result<(), StatusCode>

Closes the session and deletes all subscriptions

Returns

Ok(()) - if the session was closed

Err(StatusCode) - Status code reason for failure

pub fn subscription_state(&self) -> Arc<RwLock<SubscriptionState>>

Returns the subscription state object