opcua::client::prelude

Trait DiscoveryService

pub trait DiscoveryService: Service {
    // Required methods
    fn find_servers<T>(
        &self,
        endpoint_url: T,
    ) -> Result<Vec<ApplicationDescription>, StatusCode>
       where T: Into<UAString>;
    fn get_endpoints(&self) -> Result<Vec<EndpointDescription>, StatusCode>;
    fn register_server(
        &self,
        server: RegisteredServer,
    ) -> Result<(), StatusCode>;
}

Discovery Service set

Required Methods

fn find_servers<T>( &self, endpoint_url: T, ) -> Result<Vec<ApplicationDescription>, StatusCode>
where T: Into<UAString>,

Sends a FindServersRequest to the server denoted by the discovery url.

See OPC UA Part 4 - Services 5.4.2 for complete description of the service and error responses.

Arguments

endpoint_url - The network address that the Client used to access the Discovery Endpoint.

Returns

Ok(Vec<ApplicationDescription>) - A list of ApplicationDescription that meet criteria specified in the request.

Err(StatusCode) - Request failed, status code is the reason for failure

fn get_endpoints(&self) -> Result<Vec<EndpointDescription>, StatusCode>

Obtain the list of endpoints supported by the server by sending it a GetEndpointsRequest.

See OPC UA Part 4 - Services 5.4.4 for complete description of the service and error responses.

Returns

Ok(Vec<EndpointDescription>) - A list of endpoints supported by the server

Err(StatusCode) - Request failed, status code is the reason for failure

fn register_server(&self, server: RegisteredServer) -> Result<(), StatusCode>

This function is used by servers that wish to register themselves with a discovery server. i.e. one server is the client to another server. The server sends a RegisterServerRequest to the discovery server to register itself. Servers are expected to re-register themselves periodically with the discovery server, with a maximum of 10 minute intervals.

See OPC UA Part 4 - Services 5.4.5 for complete description of the service and error responses.

Arguments

server - The server to register

Returns

Ok(()) - Success

Err(StatusCode) - Request failed, status code is the reason for failure

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.