opcua::client::prelude

Trait MonitoredItemService

pub trait MonitoredItemService: Service {
    // Required methods
    fn create_monitored_items(
        &self,
        subscription_id: u32,
        timestamps_to_return: TimestampsToReturn,
        items_to_create: &[MonitoredItemCreateRequest],
    ) -> Result<Vec<MonitoredItemCreateResult>, StatusCode>;
    fn modify_monitored_items(
        &self,
        subscription_id: u32,
        timestamps_to_return: TimestampsToReturn,
        items_to_modify: &[MonitoredItemModifyRequest],
    ) -> Result<Vec<MonitoredItemModifyResult>, StatusCode>;
    fn set_monitoring_mode(
        &self,
        subscription_id: u32,
        monitoring_mode: MonitoringMode,
        monitored_item_ids: &[u32],
    ) -> Result<Vec<StatusCode>, StatusCode>;
    fn set_triggering(
        &self,
        subscription_id: u32,
        triggering_item_id: u32,
        links_to_add: &[u32],
        links_to_remove: &[u32],
    ) -> Result<(Option<Vec<StatusCode>>, Option<Vec<StatusCode>>), StatusCode>;
    fn delete_monitored_items(
        &self,
        subscription_id: u32,
        items_to_delete: &[u32],
    ) -> Result<Vec<StatusCode>, StatusCode>;
}

Required Methods

fn create_monitored_items( &self, subscription_id: u32, timestamps_to_return: TimestampsToReturn, items_to_create: &[MonitoredItemCreateRequest], ) -> Result<Vec<MonitoredItemCreateResult>, StatusCode>

Creates monitored items on a subscription by sending a CreateMonitoredItemsRequest to the server.

See OPC UA Part 4 - Services 5.12.2 for complete description of the service and error responses.

Arguments

subscription_id - The Server-assigned identifier for the Subscription that will report Notifications for this MonitoredItem

timestamps_to_return - An enumeration that specifies the timestamp Attributes to be transmitted for each MonitoredItem.

items_to_create - A list of MonitoredItemCreateRequest to be created and assigned to the specified Subscription.

Returns

Ok(Vec<MonitoredItemCreateResult>) - A list of MonitoredItemCreateResult corresponding to the items to create. The size and order of the list matches the size and order of the items_to_create request parameter.

Err(StatusCode) - Status code reason for failure

fn modify_monitored_items( &self, subscription_id: u32, timestamps_to_return: TimestampsToReturn, items_to_modify: &[MonitoredItemModifyRequest], ) -> Result<Vec<MonitoredItemModifyResult>, StatusCode>

Modifies monitored items on a subscription by sending a ModifyMonitoredItemsRequest to the server.

See OPC UA Part 4 - Services 5.12.3 for complete description of the service and error responses.

Arguments

subscription_id - The Server-assigned identifier for the Subscription that will report Notifications for this MonitoredItem.

timestamps_to_return - An enumeration that specifies the timestamp Attributes to be transmitted for each MonitoredItem.

items_to_modify - The list of MonitoredItemModifyRequest to modify.

Returns

Ok(Vec<MonitoredItemModifyResult>) - A list of MonitoredItemModifyResult corresponding to the MonitoredItems to modify. The size and order of the list matches the size and order of the items_to_modify request parameter.

Err(StatusCode) - Status code reason for failure.

fn set_monitoring_mode( &self, subscription_id: u32, monitoring_mode: MonitoringMode, monitored_item_ids: &[u32], ) -> Result<Vec<StatusCode>, StatusCode>

Sets the monitoring mode on one or more monitored items by sending a SetMonitoringModeRequest to the server.

See OPC UA Part 4 - Services 5.12.4 for complete description of the service and error responses.

Arguments

subscription_id - the subscription identifier containing the monitored items to be modified.

monitoring_mode - the monitored mode to apply to the monitored items

monitored_item_ids - the monitored items to be modified

Returns

Ok(Vec<StatusCode>) - Individual result for each monitored item.

Err(StatusCode) - Status code reason for failure.

fn set_triggering( &self, subscription_id: u32, triggering_item_id: u32, links_to_add: &[u32], links_to_remove: &[u32], ) -> Result<(Option<Vec<StatusCode>>, Option<Vec<StatusCode>>), StatusCode>

Sets a monitored item so it becomes the trigger that causes other monitored items to send change events in the same update. Sends a SetTriggeringRequest to the server. Note that items_to_remove is applied before items_to_add.

See OPC UA Part 4 - Services 5.12.5 for complete description of the service and error responses.

Arguments

subscription_id - the subscription identifier containing the monitored item to be used as the trigger.

monitored_item_id - the monitored item that is the trigger.

links_to_add - zero or more items to be added to the monitored item’s triggering list.

items_to_remove - zero or more items to be removed from the monitored item’s triggering list.

Returns

Ok((Option<Vec<StatusCode>>, Option<Vec<StatusCode>>)) - Individual result for each item added / removed for the SetTriggering call.

Err(StatusCode) - Status code reason for failure.

fn delete_monitored_items( &self, subscription_id: u32, items_to_delete: &[u32], ) -> Result<Vec<StatusCode>, StatusCode>

Deletes monitored items from a subscription by sending a DeleteMonitoredItemsRequest to the server.

See OPC UA Part 4 - Services 5.12.6 for complete description of the service and error responses.

Arguments

subscription_id - The Server-assigned identifier for the Subscription that will report Notifications for this MonitoredItem.

items_to_delete - List of Server-assigned ids for the MonitoredItems to be deleted.

Returns

Ok(Vec<StatusCode>) - List of StatusCodes for the MonitoredItems to delete. The size and order of the list matches the size and order of the items_to_delete request parameter.

Err(StatusCode) - Status code reason for failure.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.