opcua::client::prelude

Trait NodeManagementService

pub trait NodeManagementService: Service {
    // Required methods
    fn add_nodes(
        &self,
        nodes_to_add: &[AddNodesItem],
    ) -> Result<Vec<AddNodesResult>, StatusCode>;
    fn add_references(
        &self,
        references_to_add: &[AddReferencesItem],
    ) -> Result<Vec<StatusCode>, StatusCode>;
    fn delete_nodes(
        &self,
        nodes_to_delete: &[DeleteNodesItem],
    ) -> Result<Vec<StatusCode>, StatusCode>;
    fn delete_references(
        &self,
        references_to_delete: &[DeleteReferencesItem],
    ) -> Result<Vec<StatusCode>, StatusCode>;
}

NodeManagement Service set

Required Methods

fn add_nodes( &self, nodes_to_add: &[AddNodesItem], ) -> Result<Vec<AddNodesResult>, StatusCode>

Add nodes by sending a AddNodesRequest to the server.

See OPC UA Part 4 - Services 5.7.2 for complete description of the service and error responses.

Arguments

nodes_to_add - A list of AddNodesItem to be added to the server.

Returns

Ok(Vec<AddNodesResult>) - A list of AddNodesResult corresponding to each add node operation.

Err(StatusCode) - Status code reason for failure.

fn add_references( &self, references_to_add: &[AddReferencesItem], ) -> Result<Vec<StatusCode>, StatusCode>

Add references by sending a AddReferencesRequest to the server.

See OPC UA Part 4 - Services 5.7.3 for complete description of the service and error responses.

Arguments

references_to_add - A list of AddReferencesItem to be sent to the server.

Returns

Ok(Vec<StatusCode>) - A list of StatusCode corresponding to each add reference operation.

Err(StatusCode) - Status code reason for failure.

fn delete_nodes( &self, nodes_to_delete: &[DeleteNodesItem], ) -> Result<Vec<StatusCode>, StatusCode>

Delete nodes by sending a DeleteNodesRequest to the server.

See OPC UA Part 4 - Services 5.7.4 for complete description of the service and error responses.

Arguments

nodes_to_delete - A list of DeleteNodesItem to be sent to the server.

Returns

Ok(Vec<StatusCode>) - A list of StatusCode corresponding to each delete node operation.

Err(StatusCode) - Status code reason for failure.

fn delete_references( &self, references_to_delete: &[DeleteReferencesItem], ) -> Result<Vec<StatusCode>, StatusCode>

Delete references by sending a DeleteReferencesRequest to the server.

See OPC UA Part 4 - Services 5.7.5 for complete description of the service and error responses.

Arguments

nodes_to_delete - A list of DeleteReferencesItem to be sent to the server.

Returns

Ok(Vec<StatusCode>) - A list of StatusCode corresponding to each delete node operation.

Err(StatusCode) - Status code reason for failure.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.