opcua::client::prelude

Trait OnSessionClosed

pub trait OnSessionClosed {
    // Required method
    fn on_session_closed(&mut self, status_code: StatusCode);
}

The OnSessionClosed trait can be used to register on a session and called to notify the client that the session has closed.

Required Methods

fn on_session_closed(&mut self, status_code: StatusCode)

Called when the connection closed (in addition to a status change event). The status code should be checked to see if the closure was a graceful terminate (Good), or the result of a network or protocol error.

If no session retry policy has been created for the client session, the server implementation might choose to reconnect in response to a bad status code by itself, however it should avoid retrying too quickly or indefinitely in case the error is permanent.