opcua::client::prelude

Trait OnSubscriptionNotification

pub trait OnSubscriptionNotification {
    // Provided methods
    fn on_data_change(&mut self, _data_change_items: &[&MonitoredItem]) { ... }
    fn on_event(&mut self, _events: &EventNotificationList) { ... }
}

The OnSubscriptionNotification trait is the callback registered along with a new subscription to receive subscription notification callbacks.

Unless your subscription contains a mix of items which are monitoring data and events you probably only need to implement either data_change(), or event() and leave the default, no-op implementation for the other.

There are concrete implementations of this trait in DataChangeCallback and EventCallback.

Provided Methods

fn on_data_change(&mut self, _data_change_items: &[&MonitoredItem])

Called by the subscription after a DataChangeNotification. The default implementation does nothing.

fn on_event(&mut self, _events: &EventNotificationList)

Called by the subscription after a EventNotificationList. The notifications contained within are individual EventFieldList structs filled from the select clause criteria from when the event was constructed. The default implementation does nothing.