opcua::client::prelude

Trait SessionService

pub trait SessionService: Service {
    // Required methods
    fn create_session(&self) -> Result<NodeId, StatusCode>;
    fn activate_session(&self) -> Result<(), StatusCode>;
    fn cancel(&self, request_handle: IntegerId) -> Result<u32, StatusCode>;
}

Session Service set

Required Methods

fn create_session(&self) -> Result<NodeId, StatusCode>

Sends a CreateSessionRequest to the server, returning the session id of the created session. Internally, the session will store the authentication token which is used for requests subsequent to this call.

See OPC UA Part 4 - Services 5.6.2 for complete description of the service and error responses.

Returns

Ok(NodeId) - Success, session id

Err(StatusCode) - Request failed, status code is the reason for failure

fn activate_session(&self) -> Result<(), StatusCode>

Sends an ActivateSessionRequest to the server to activate this session

See OPC UA Part 4 - Services 5.6.3 for complete description of the service and error responses.

Returns

Ok(()) - Success

Err(StatusCode) - Request failed, status code is the reason for failure

fn cancel(&self, request_handle: IntegerId) -> Result<u32, StatusCode>

Cancels an outstanding service request by sending a CancelRequest to the server.

See OPC UA Part 4 - Services 5.6.5 for complete description of the service and error responses.

Arguments

request_handle - Handle to the outstanding request to be cancelled.

Returns

Ok(u32) - Success, number of cancelled requests

Err(StatusCode) - Request failed, status code is the reason for failure

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.