opcua::core

Module comms

Contains all code related to sending / receiving messages from a transport and turning those messages into and out of chunks.

Modules

chunker

Contains code for turning messages into chunks and chunks into messages.

message_chunk

A message chunk is a message or a portion of a message, optionally encrypted & signed, which has been split for transmission.

message_chunk_info

message_writer

prelude

secure_channel

security_header

tcp_codec

The codec is an implementation of a tokio Encoder/Decoder which can be used to read data from the socket in terms of frames which in our case are any of the following:

tcp_types

Contains the implementation of various UA over TCP types.

url

Provides functions for parsing Urls from strings.