opcua::core::comms

Module message_chunk

A message chunk is a message or a portion of a message, optionally encrypted & signed, which has been split for transmission.

Structs

MessageChunk

A chunk holds a message or a portion of a message, if the message has been split into multiple chunks. The chunk’s data may be signed and encrypted. To extract the message requires all the chunks to be available in sequence so they can be formed back into the message.

MessageChunkHeader

Enums

MessageChunkType

MessageIsFinalType

Constants

MESSAGE_CHUNK_HEADER_SIZE

The size of a chunk header, used by several places