opcua::core::comms

Module tcp_codec

The codec is an implementation of a tokio Encoder/Decoder which can be used to read data from the socket in terms of frames which in our case are any of the following:

HEL - Hello message

ACK - Acknowledge message

ERR - Error message

MSG - Message chunk

OPN - Open Secure Channel message

CLO - Close Secure Channel message

Structs

TcpCodec

Implements a tokio codec that as close as possible, allows incoming data to be transformed into OPC UA message chunks with no intermediate buffers. Chunks are subsequently transformed into messages so there is still some buffers within message chunks, but not at the raw socket level.

Enums

Message