opcua::core::comms::tcp_codec

Struct TcpCodec

pub struct TcpCodec { /* private fields */ }

Implements a tokio codec that as close as possible, allows incoming data to be transformed into OPC UA message chunks with no intermediate buffers. Chunks are subsequently transformed into messages so there is still some buffers within message chunks, but not at the raw socket level.

Implementations

impl TcpCodec

pub fn new(decoding_options: DecodingOptions) -> TcpCodec

Constructs a new TcpCodec. The abort flag is set to terminate the codec even while it is waiting for a frame to arrive.