opcua::core::comms

Module url

Provides functions for parsing Urls from strings.

Constants

OPC_TCP_SCHEME

Functions

hostname_from_url

hostname_port_from_url

is_opc_ua_binary_url

is_valid_opc_ua_url

server_url_from_endpoint_url

Takes an endpoint url and strips off the path and args to leave just the protocol, host & port.

url_matches_except_host

Test if the two urls match except for the hostname. Can be used by a server whose endpoint doesn’t exactly match the incoming connection, e.g. 127.0.0.1 vs localhost.

url_with_replaced_hostname

Replace the hostname in the supplied url and return a new url