opcua::core::comms::url

Function url_matches_except_host

pub fn url_matches_except_host(url1: &str, url2: &str) -> bool

Test if the two urls match except for the hostname. Can be used by a server whose endpoint doesn’t exactly match the incoming connection, e.g. 127.0.0.1 vs localhost.