opcua

Module crypto

Crypto related functionality. It is used for establishing trust between a client and server via certificate exchange and validation. It also used for encrypting / decrypting messages and signing messages.

Re-exports

pub use aeskey::*;

pub use certificate_store::*;

pub use hash::*;

pub use pkey::*;

pub use security_policy::*;

pub use thumbprint::*;

pub use user_identity::*;

pub use x509::*;

Modules

aeskey

Symmetric encryption / decryption wrapper.

certificate_store

The certificate store holds and retrieves private keys and certificates from disk. It is responsible for checking certificates supplied by the remote end to see if they are valid and trusted or not.

hash

Hashing functions used for producing and verifying digital signatures

pkey

Asymmetric encryption / decryption, signing / verification wrapper.

random

Module contains functions for creating cryptographically strong random bytes.

security_policy

Security policy is the symmetric, asymmetric encryption / decryption + signing / verification algorithms to use and enforce for the current session.

thumbprint

Functionality for holding a message digest.

user_identity

Functions related to encrypting / decrypting passwords in a UserNameIdentityToken.

x509

Structs

HostnameError

Constants

SHA1_SIZE

SHA256_SIZE

Functions

create_signature_data

Creates a SignatureData object by signing the supplied certificate and nonce with a pkey

hostname

Returns this computer’s hostname

verify_signature_data

Verifies that the supplied signature data was produced by the signing cert. The contained cert and nonce are supplied so the signature can be verified against the expected data.