opcua::crypto

Module hash

Hashing functions used for producing and verifying digital signatures

Functions

hmac_sha1

hmac_sha256

p_sha

Pseudo random P_SHA implementation for creating pseudo random range of bytes from an input

verify_hmac_sha1

Verify that the HMAC for the data block matches the supplied signature

verify_hmac_sha256

Verify that the HMAC for the data block matches the supplied signature