opcua::crypto::pkey

Type Alias PrivateKey

pub type PrivateKey = PKey<Private>;

Aliased Type

pub struct PrivateKey { /* private fields */ }

Implementations

impl PrivateKey

pub fn new(bit_length: u32) -> PrivateKey

pub fn wrap_private_key(pkey: PKey<Private>) -> PrivateKey

pub fn from_pem(pem: &[u8]) -> Result<PrivateKey, PKeyError>

pub fn private_key_to_pem(&self) -> Result<Vec<u8>, PKeyError>

pub fn sign_sha1( &self, data: &[u8], signature: &mut [u8], ) -> Result<usize, StatusCode>

Signs the data using RSA-SHA1

pub fn sign_sha256( &self, data: &[u8], signature: &mut [u8], ) -> Result<usize, StatusCode>

Signs the data using RSA-SHA256

pub fn sign_sha256_pss( &self, data: &[u8], signature: &mut [u8], ) -> Result<usize, StatusCode>

Signs the data using RSA-SHA256-PSS

pub fn private_decrypt( &self, src: &[u8], dst: &mut [u8], padding: RsaPadding, ) -> Result<usize, PKeyError>

Decrypts data in src to dst using the specified padding and returning the size of the decrypted data in bytes or an error.