opcua::crypto::pkey

Type Alias PublicKey

pub type PublicKey = PKey<Public>;

A public key

Aliased Type

pub struct PublicKey { /* private fields */ }

Implementations

impl PublicKey

pub fn wrap_public_key(pkey: PKey<Public>) -> PublicKey

pub fn verify_sha1( &self, data: &[u8], signature: &[u8], ) -> Result<bool, StatusCode>

Verifies the data using RSA-SHA1

pub fn verify_sha256( &self, data: &[u8], signature: &[u8], ) -> Result<bool, StatusCode>

Verifies the data using RSA-SHA256

pub fn verify_sha256_pss( &self, data: &[u8], signature: &[u8], ) -> Result<bool, StatusCode>

Verifies the data using RSA-SHA256-PSS

pub fn public_encrypt( &self, src: &[u8], dst: &mut [u8], padding: RsaPadding, ) -> Result<usize, PKeyError>

Encrypts data from src to dst using the specified padding and returns the size of encrypted data in bytes or an error.