opcua::crypto::x509

Struct X509

pub struct X509 { /* private fields */ }

This is a wrapper around the OpenSSL X509 cert

Implementations

impl X509

pub fn from_der(der: &[u8]) -> Result<Self, X509Error>

pub fn cert_and_pkey(x509_data: &X509Data) -> Result<(Self, PrivateKey), String>

Creates a self-signed X509v3 certificate and public/private key from the supplied creation args. The certificate identifies an instance of the application running on a host as well as the public key. The PKey holds the corresponding public/private key. Note that if the pkey is stored by cert store, then only the private key will be written. The public key is only ever stored with the cert.

See Part 6 Table 23 for full set of requirements

In particular, application instance cert requires subjectAltName to specify alternate hostnames / ip addresses that the host runs on.

pub fn from_pkey( pkey: &PrivateKey, x509_data: &X509Data, ) -> Result<Self, String>

pub fn from_byte_string(data: &ByteString) -> Result<X509, StatusCode>

pub fn as_byte_string(&self) -> ByteString

Returns a ByteString representation of the cert which is DER encoded form of X509v3

pub fn public_key(&self) -> Result<PublicKey, StatusCode>

pub fn key_length(&self) -> Result<usize, X509Error>

Returns the key length in bits (if possible)

pub fn subject_name(&self) -> String

pub fn common_name(&self) -> Result<String, X509Error>

Gets the common name out of the cert

pub fn is_time_valid(&self, now: &DateTime<Utc>) -> StatusCode

Tests if the certificate is valid for the supplied time using the not before and not after values on the cert.

pub fn is_hostname_valid(&self, hostname: &str) -> StatusCode

Tests if the supplied hostname matches any of the dns alt subject name entries on the cert

pub fn is_application_uri_valid(&self, application_uri: &str) -> StatusCode

Tests if the supplied application uri matches the uri alt subject name entry on the cert

pub fn thumbprint(&self) -> Thumbprint

OPC UA Part 6 MessageChunk structure

The thumbprint is the SHA1 digest of the DER form of the certificate. The hash is 160 bits (20 bytes) in length and is sent in some secure conversation headers.

The thumbprint might be used by the server / client for look-up purposes.

pub fn not_before(&self) -> Result<DateTime<Utc>, X509Error>

Turn the Asn1 values into useful portable types

pub fn not_after(&self) -> Result<DateTime<Utc>, X509Error>

Turn the Asn1 values into useful portable types

pub fn to_der(&self) -> Result<Vec<u8>, X509Error>