opcua::server

Module address_space

Provides functionality to create an address space, find nodes, add nodes, change attributes and values on nodes.

Re-exports

pub use self::address_space::AddressSpace;

Modules

address_space

Implementation of AddressSpace.

base

data_type

Contains the implementation of Method and MethodBuilder.

method

Contains the implementation of Method and MethodBuilder.

node

object

Contains the implementation of Object and ObjectBuilder.

object_type

Contains the implementation of ObjectType and ObjectTypeBuilder.

reference_type

Contains the implementation of ReferenceType and ReferenceTypeBuilder.

references

relative_path

types

variable

Contains the implementation of Variable and VariableBuilder.

variable_type

Contains the implementation of VariableType and VariableTypeBuilder.

view

Contains the implementation of View and ViewBuilder.

Structs

AccessLevel

AttrFnGetter

An implementation of attribute getter that can be easily constructed from a mutable function

AttrFnSetter

An implementation of attribute setter that can be easily constructed using a mutable function

EventNotifier

UserAccessLevel