opcua::server::address_space::node

Trait NodeBase

pub trait NodeBase {
    // Required methods
    fn node_class(&self) -> NodeClass;
    fn node_id(&self) -> NodeId;
    fn browse_name(&self) -> QualifiedName;
    fn display_name(&self) -> LocalizedText;
    fn set_display_name(&mut self, display_name: LocalizedText);
    fn description(&self) -> Option<LocalizedText>;
    fn set_description(&mut self, description: LocalizedText);
    fn write_mask(&self) -> Option<WriteMask>;
    fn set_write_mask(&mut self, write_mask: WriteMask);
    fn user_write_mask(&self) -> Option<WriteMask>;
    fn set_user_write_mask(&mut self, write_mask: WriteMask);
}

Implemented within a macro for all Node types. Functions that return a result in an Option do so because the attribute is optional and not necessarily there.

Required Methods

fn node_class(&self) -> NodeClass

Returns the node class - Object, ObjectType, Method, DataType, ReferenceType, Variable, VariableType or View

fn node_id(&self) -> NodeId

Returns the node’s NodeId

fn browse_name(&self) -> QualifiedName

Returns the node’s browse name

fn display_name(&self) -> LocalizedText

Returns the node’s display name

fn set_display_name(&mut self, display_name: LocalizedText)

Sets the node’s display name

fn description(&self) -> Option<LocalizedText>

fn set_description(&mut self, description: LocalizedText)

fn write_mask(&self) -> Option<WriteMask>

fn set_write_mask(&mut self, write_mask: WriteMask)

fn user_write_mask(&self) -> Option<WriteMask>

fn set_user_write_mask(&mut self, write_mask: WriteMask)