opcua::server::address_space::view

Struct ViewBuilder

pub struct ViewBuilder { /* private fields */ }

A builder for constructing a node of same name. This can be used as an easy way to create a node and the references it has to another node in a simple fashion.

Implementations

impl ViewBuilder

pub fn new<T, S>(node_id: &NodeId, browse_name: T, display_name: S) -> Self
where T: Into<QualifiedName>, S: Into<LocalizedText>,

Creates a builder for a node. All nodes are required to su

pub fn get_node_id(&self) -> NodeId

pub fn is_valid(&self) -> bool

Tests that the builder is in a valid state to build or insert the node.

pub fn description<V>(self, description: V) -> Self
where V: Into<LocalizedText>,

Sets the description of the node

pub fn reference<T>( self, node_id: T, reference_type_id: ReferenceTypeId, reference_direction: ReferenceDirection, ) -> Self
where T: Into<NodeId>,

Adds a reference to the node

pub fn organizes<T>(self, organizes_id: T) -> Self
where T: Into<NodeId>,

Indicates this node organizes another node by its id.

pub fn organized_by<T>(self, organized_by_id: T) -> Self
where T: Into<NodeId>,

Indicates this node is organised by another node by its id

pub fn build(self) -> View

Yields a built node. This function will panic if the node is invalid. Note that calling this function discards any references for the node, so there is no purpose in adding references if you intend to call this method.

pub fn insert(self, address_space: &mut AddressSpace) -> bool

Inserts the node into the address space, including references. This function will panic if the node is in an invalid state.

impl ViewBuilder

pub fn generates_event<T>(self, event_type: T) -> Self
where T: Into<NodeId>,

impl ViewBuilder

pub fn contains_no_loops(self, contains_no_loops: bool) -> Self

pub fn event_notifier(self, event_notifier: EventNotifier) -> Self