opcua::server::comms::transport

Trait Transport

pub trait Transport {
    // Required methods
    fn state(&self) -> TransportState;
    fn finish(&mut self, status_code: StatusCode);
    fn client_address(&self) -> Option<SocketAddr>;
    fn session_manager(&self) -> Arc<RwLock<SessionManager>>;

    // Provided methods
    fn has_received_hello(&self) -> bool { ... }
    fn is_finished(&self) -> bool { ... }
}

Represents a transport layer, the thing responsible for maintaining an open channel and transferring data between the server and the client.

Required Methods

fn state(&self) -> TransportState

fn finish(&mut self, status_code: StatusCode)

Terminate the session and put the connection in a finished state

fn client_address(&self) -> Option<SocketAddr>

Returns the address of the client (peer) of this connection

fn session_manager(&self) -> Arc<RwLock<SessionManager>>

Returns the session map for the connection

Provided Methods

fn has_received_hello(&self) -> bool

fn is_finished(&self) -> bool