opcua::server::config

Struct ServerEndpoint

pub struct ServerEndpoint {
    pub path: String,
    pub security_policy: String,
    pub security_mode: String,
    pub security_level: u8,
    pub password_security_policy: Option<String>,
    pub user_token_ids: BTreeSet<String>,
}

Fields
path: String

Endpoint path

security_policy: String

Security policy

security_mode: String

Security mode

security_level: u8

Security level, higher being more secure

password_security_policy: Option<String>

Password security policy when a client supplies a user name identity token

user_token_ids: BTreeSet<String>

User tokens

Implementations

impl ServerEndpoint

pub fn new<T>( path: T, security_policy: SecurityPolicy, security_mode: MessageSecurityMode, user_token_ids: &[String], ) -> Self
where T: Into<String>,

pub fn new_none<T>(path: T, user_token_ids: &[String]) -> Self
where T: Into<String>,

pub fn new_basic128rsa15_sign<T>(path: T, user_token_ids: &[String]) -> Self
where T: Into<String>,

pub fn new_basic128rsa15_sign_encrypt<T>( path: T, user_token_ids: &[String], ) -> Self
where T: Into<String>,

pub fn new_basic256_sign<T>(path: T, user_token_ids: &[String]) -> Self
where T: Into<String>,

pub fn new_basic256_sign_encrypt<T>(path: T, user_token_ids: &[String]) -> Self
where T: Into<String>,

pub fn new_basic256sha256_sign<T>(path: T, user_token_ids: &[String]) -> Self
where T: Into<String>,

pub fn new_basic256sha256_sign_encrypt<T>( path: T, user_token_ids: &[String], ) -> Self
where T: Into<String>,

pub fn new_aes128_sha256_rsaoaep_sign<T>( path: T, user_token_ids: &[String], ) -> Self
where T: Into<String>,

pub fn new_aes128_sha256_rsaoaep_sign_encrypt<T>( path: T, user_token_ids: &[String], ) -> Self
where T: Into<String>,

pub fn new_aes256_sha256_rsapss_sign<T>( path: T, user_token_ids: &[String], ) -> Self
where T: Into<String>,

pub fn new_aes256_sha256_rsapss_sign_encrypt<T>( path: T, user_token_ids: &[String], ) -> Self
where T: Into<String>,

pub fn is_valid( &self, id: &str, user_tokens: &BTreeMap<String, ServerUserToken>, ) -> bool

pub fn security_policy(&self) -> SecurityPolicy

pub fn message_security_mode(&self) -> MessageSecurityMode

pub fn endpoint_url(&self, base_endpoint: &str) -> String

pub fn password_security_policy(&self) -> SecurityPolicy

Returns the effective password security policy for the endpoint. This is the explicitly set password security policy, or just the regular security policy.

pub fn supports_anonymous(&self) -> bool

Test if the endpoint supports anonymous users

pub fn supports_user_pass( &self, server_tokens: &BTreeMap<String, ServerUserToken>, ) -> bool

Tests if this endpoint supports user pass tokens. It does this by looking to see if any of the users allowed to access this endpoint are user pass users.

pub fn supports_x509( &self, server_tokens: &BTreeMap<String, ServerUserToken>, ) -> bool

Tests if this endpoint supports x509 tokens. It does this by looking to see if any of the users allowed to access this endpoint are x509 users.

pub fn supports_user_token_id(&self, id: &str) -> bool