opcua::server

Module constants

Provides constants that govern the internal workings of the server implementation.

Constants

DEFAULT_DISCOVERY_SERVER_URL

Default, well known address for TCP discovery server

DEFAULT_HELLO_TIMEOUT_SECONDS

The default hello timeout period in seconds

DEFAULT_KEEP_ALIVE_COUNT

Default keep alive count

DEFAULT_MAX_MONITORED_ITEMS_PER_SUB

Default maximum number of monitored items per subscription

DEFAULT_MAX_SUBSCRIPTIONS

Default maximum number of subscriptions in a session

DEFAULT_MONITORED_ITEM_CAPACITY

The default size of preallocated vecs of monitored items per subscription

DEFAULT_RUST_OPC_UA_SERVER_PORT

Default OPC UA server port for this implementation

HELLO_TIMEOUT_POLL_MS

Interval to check for HELLO timeout in millis. This can be fairly coarse because it’s not something that requires huge accuracy.

MAX_BROWSE_CONTINUATION_POINTS

Maximum browse continuation points

MAX_DATA_CHANGE_QUEUE_SIZE

Maximum data change queue allowed by clients on monitored items

MAX_HISTORY_CONTINUATION_POINTS

Maximum history continuation points

MAX_KEEP_ALIVE_COUNT

Maximum keep alive count

MAX_MONITORED_ITEMS_PER_CALL

MAX_NODES_PER_BROWSE

MAX_NODES_PER_HISTORY_READ_DATA

MAX_NODES_PER_HISTORY_READ_EVENTS

MAX_NODES_PER_HISTORY_UPDATE_DATA

MAX_NODES_PER_HISTORY_UPDATE_EVENTS

MAX_NODES_PER_METHOD_CALL

MAX_NODES_PER_NODE_MANAGEMENT

Maximum number of nodes / references per node manaument operation

MAX_NODES_PER_READ

MAX_NODES_PER_REGISTER_NODES

MAX_NODES_PER_TRANSLATE_BROWSE_PATHS_TO_NODE_IDS

Maximum number of nodes in a TranslateBrowsePathsToNodeIdsRequest

MAX_NODES_PER_WRITE

MAX_QUERY_CONTINUATION_POINTS

Maximum query continuation points

MAX_REQUEST_MESSAGE_SIZE

Maximum size in bytes that a request message is allowed to be

MAX_SESSION_TIMEOUT

Maximum time in MS that a session can be inactive before a timeout

MIN_PUBLISHING_INTERVAL

Minimum publishing interval for subscriptions

MIN_SAMPLING_INTERVAL

Minimum sampling interval on monitored items

SUBSCRIPTION_TIMER_RATE_MS

The polling interval in millis on subscriptions and monitored items. The more fine-grained this is, the more often subscriptions will be checked for changes. The minimum publish interval cannot be less than this.