opcua::server::events

Module event

Contains functions for generating events and adding them to the address space of the server.

Structs

BaseEventType

This corresponds to BaseEventType definition in OPC UA Part 5

Traits

Event

Events can implement this to populate themselves into the address space

Functions

events_for_object

Searches for events of the specified event type which reference the source object

filter_events

Attempts to find events that were emitted by the source object based upon a time predicate

purge_events