opcua::server::historical

Trait HistoricalDataProvider

pub trait HistoricalDataProvider {
    // Provided methods
    fn read_raw_modified_details(
        &self,
        _address_space: Arc<RwLock<AddressSpace>>,
        _request: ReadRawModifiedDetails,
        _timestamps_to_return: TimestampsToReturn,
        _release_continuation_points: bool,
        _nodes_to_read: &[HistoryReadValueId],
    ) -> Result<Vec<HistoryReadResult>, StatusCode> { ... }
    fn read_processed_details(
        &self,
        _address_space: Arc<RwLock<AddressSpace>>,
        _request: ReadProcessedDetails,
        _timestamps_to_return: TimestampsToReturn,
        _release_continuation_points: bool,
        _nodes_to_read: &[HistoryReadValueId],
    ) -> Result<Vec<HistoryReadResult>, StatusCode> { ... }
    fn read_at_time_details(
        &self,
        _address_space: Arc<RwLock<AddressSpace>>,
        _request: ReadAtTimeDetails,
        _timestamps_to_return: TimestampsToReturn,
        _release_continuation_points: bool,
        _nodes_to_read: &[HistoryReadValueId],
    ) -> Result<Vec<HistoryReadResult>, StatusCode> { ... }
    fn update_data_details(
        &self,
        _address_space: Arc<RwLock<AddressSpace>>,
        _request: UpdateDataDetails,
    ) -> Result<Vec<StatusCode>, StatusCode> { ... }
    fn update_structure_data_details(
        &self,
        _address_space: Arc<RwLock<AddressSpace>>,
        _request: UpdateStructureDataDetails,
    ) -> Result<Vec<StatusCode>, StatusCode> { ... }
    fn delete_raw_modified_details(
        &self,
        _address_space: Arc<RwLock<AddressSpace>>,
        _request: DeleteRawModifiedDetails,
    ) -> Result<Vec<StatusCode>, StatusCode> { ... }
    fn delete_at_time_details(
        &self,
        _address_space: Arc<RwLock<AddressSpace>>,
        _request: DeleteAtTimeDetails,
    ) -> Result<Vec<StatusCode>, StatusCode> { ... }
}

The HistoricalEventProvider trait provides the function stubs that a server will call to process historical data operations. The implementor of this trait may provide their own implementation as many functions as they desire leaving the remainder as stubs.

IMPORTANT NOTE: This trait is currently synchronous and may change in the future to some other form. In the meantime it means if you are doing lengthy reads then use continuation points to spawn a thread for that activity. Updates and deletes should be spawned on separate threads if they are lengthy operations.

Provided Methods

fn read_raw_modified_details( &self, _address_space: Arc<RwLock<AddressSpace>>, _request: ReadRawModifiedDetails, _timestamps_to_return: TimestampsToReturn, _release_continuation_points: bool, _nodes_to_read: &[HistoryReadValueId], ) -> Result<Vec<HistoryReadResult>, StatusCode>

Note: Function returns an HistoryRawData enum containing either a HistoryData for a read raw action or a HistoryModifiedData for a read modified action.

fn read_processed_details( &self, _address_space: Arc<RwLock<AddressSpace>>, _request: ReadProcessedDetails, _timestamps_to_return: TimestampsToReturn, _release_continuation_points: bool, _nodes_to_read: &[HistoryReadValueId], ) -> Result<Vec<HistoryReadResult>, StatusCode>

fn read_at_time_details( &self, _address_space: Arc<RwLock<AddressSpace>>, _request: ReadAtTimeDetails, _timestamps_to_return: TimestampsToReturn, _release_continuation_points: bool, _nodes_to_read: &[HistoryReadValueId], ) -> Result<Vec<HistoryReadResult>, StatusCode>

fn update_data_details( &self, _address_space: Arc<RwLock<AddressSpace>>, _request: UpdateDataDetails, ) -> Result<Vec<StatusCode>, StatusCode>

fn update_structure_data_details( &self, _address_space: Arc<RwLock<AddressSpace>>, _request: UpdateStructureDataDetails, ) -> Result<Vec<StatusCode>, StatusCode>

fn delete_raw_modified_details( &self, _address_space: Arc<RwLock<AddressSpace>>, _request: DeleteRawModifiedDetails, ) -> Result<Vec<StatusCode>, StatusCode>

fn delete_at_time_details( &self, _address_space: Arc<RwLock<AddressSpace>>, _request: DeleteAtTimeDetails, ) -> Result<Vec<StatusCode>, StatusCode>