opcua::server::session

Struct Session

pub struct Session { /* private fields */ }

The Session is any state maintained between the client and server

Implementations

impl Session

pub fn new(server_state: Arc<RwLock<ServerState>>) -> Session

Create a Session from a Server

pub fn session_id(&self) -> &NodeId

pub fn set_activated(&mut self, activated: bool)

pub fn is_activated(&self) -> bool

pub fn is_terminated(&self) -> bool

pub fn terminated_at(&self) -> DateTimeUtc

pub fn set_terminated(&mut self)

pub fn secure_channel_id(&self) -> u32

pub fn set_secure_channel_id(&mut self, secure_channel_id: u32)

pub fn authentication_token(&self) -> &NodeId

pub fn set_authentication_token(&mut self, authentication_token: NodeId)

pub fn session_timeout(&self) -> f64

pub fn set_session_timeout(&mut self, session_timeout: f64)

pub fn set_max_request_message_size(&mut self, max_request_message_size: u32)

pub fn set_max_response_message_size(&mut self, max_response_message_size: u32)

pub fn endpoint_url(&self) -> &UAString

pub fn set_endpoint_url(&mut self, endpoint_url: UAString)

pub fn set_security_policy_uri(&mut self, security_policy_uri: &str)

pub fn set_user_identity(&mut self, user_identity: IdentityToken)

pub fn last_service_request_timestamp(&self) -> DateTimeUtc

pub fn set_last_service_request_timestamp( &mut self, last_service_request_timestamp: DateTimeUtc, )

pub fn locale_ids(&self) -> &Option<Vec<UAString>>

pub fn set_locale_ids(&mut self, locale_ids: Option<Vec<UAString>>)

pub fn client_certificate(&self) -> &Option<X509>

pub fn set_client_certificate(&mut self, client_certificate: Option<X509>)

pub fn session_nonce(&self) -> &ByteString

pub fn set_session_nonce(&mut self, session_nonce: ByteString)

pub fn session_name(&self) -> &UAString

pub fn set_session_name(&mut self, session_name: UAString)

pub fn client_user_id(&self) -> UAString

Helper function to return the client user id from the identity token or None of there is no user id

This conforms to OPC Part 5 6.4.3 ClientUserId

pub fn is_session_terminated(&self) -> bool

pub fn terminate_session(&mut self)