opcua::server::session

Struct SessionManager

pub struct SessionManager {
    pub sessions: HashMap<NodeId, Arc<RwLock<Session>>>,
    pub sessions_terminated: bool,
}

Fields
sessions: HashMap<NodeId, Arc<RwLock<Session>>>sessions_terminated: bool
Implementations

impl SessionManager

pub fn len(&self) -> usize

pub fn first(&self) -> Option<Arc<RwLock<Session>>>

pub fn sessions_terminated(&self) -> bool

pub fn clear(&mut self, address_space: Arc<RwLock<AddressSpace>>)

Puts all sessions into a terminated state, deregisters them, and clears the map

pub fn find_session_by_id( &self, session_id: &NodeId, ) -> Option<Arc<RwLock<Session>>>

Find a session by its session id and return it.

pub fn find_session_by_token( &self, authentication_token: &NodeId, ) -> Option<Arc<RwLock<Session>>>

Finds the session by its authentication token and returns it. The authentication token can be renewed so it is not used as a key.

pub fn register_session(&mut self, session: Arc<RwLock<Session>>)

Register the session in the map so it can be searched on

pub fn deregister_session( &mut self, session: Arc<RwLock<Session>>, ) -> Option<Arc<RwLock<Session>>>

Deregisters a session from the map