opcua::server

Module state

Provides server state information, such as status, configuration, running servers and so on.

Structs

ServerState

Server state is any state associated with the server as a whole that individual sessions might be interested in. That includes configuration info etc.