opcua

Module types

The OPC UA Types module contains data types and enumerations for OPC UA.

This includes:

All of the built-in data types described in OPC Part 6 Chapter 5 that are encodable.

All of the standard data types described in OPC Part 3 Chapter 8 (if not covered by 1.).

Autogenerated data types and request / responses as described in OPC Part 4.

For the built-in data types, the module provides functions

Re-exports

pub use crate::types::argument::*;

pub use crate::types::array::*;

pub use crate::types::attribute::*;

pub use crate::types::basic_types::*;

pub use crate::types::byte_string::*;

pub use crate::types::data_types::*;

pub use crate::types::data_value::*;

pub use crate::types::date_time::*;

pub use crate::types::diagnostic_info::*;

pub use crate::types::encoding::*;

pub use crate::types::expanded_node_id::*;

pub use crate::types::extension_object::*;

pub use crate::types::guid::*;

pub use crate::types::localized_text::*;

pub use crate::types::node_id::*;

pub use crate::types::node_ids::*;

pub use crate::types::numeric_range::*;

pub use crate::types::operand::*;

pub use crate::types::qualified_name::*;

pub use crate::types::request_header::*;

pub use crate::types::response_header::*;

pub use crate::types::service_types::*;

pub use crate::types::status_code::*;

pub use crate::types::string::*;

pub use crate::types::variant::*;

pub use crate::types::variant_type_id::*;

Modules

argument

array

attribute

basic_types

Contains definitions of the simple OPC UA scalar types.

byte_string

Contains the implementation of ByteString.

constants

data_types

data_value

Contains the implementation of DataValue.

date_time

Contains the implementation of DataTime.

diagnostic_info

Contains the implementation of DiagnosticInfo.

encoding

Contains the BinaryEncoder trait and helpers for reading and writing of scalar values and other primitives.

expanded_node_id

Contains the implementation of ExpandedNodeId.

extension_object

Contains the implementation of ExtensionObject.

guid

Contains the implementation of Guid.

localized_text

Contains the definition of LocalizedText.

node_id

Contains the implementation of NodeId.

node_ids

notification_message

numeric_range

Contains the implementation of NumericRange.

operand

profiles

Contains constants recognized by OPC UA clients and servers to describe various protocols and profiles used during communication and encryption.

qualified_name

Contains the definition of QualifiedName.

relative_path

Contains functions used for making relative paths from / to strings, as per OPC UA Part 4, Appendix A

request_header

response_header

service_types

status_code

Contains the hand implemented part of the StatusCode type. The other file, status_codes.rs contains the machine generated part.

string

Contains the implementation of UAString.

variant

Contains the most of the implementation of Variant. Some substantial chunks like JSON serialization are moved off into their own files due to the complexity of this functionality.

variant_json

Implements the JSON serialization functionality for Variant. OPC UA serialization over JSON is described by Part 6 of the spec in not a whole lot of detail, so this implementation tries its best to implement it as described. In some places there may be ambiguity, e.g. whether a null field corresponds to null in one of the types. In addition, serialization of JSON is so different than the binary form that it doesn’t share any code with OPC UA binary.

variant_type_id

Structs

AttributesMask

BrowseDescriptionResultMask

NodeClassMask

WriteMask