opcua::types

Module constants

Constants

MAX_ARRAY_LENGTH

Default maximum number of elements in an array

MAX_BYTE_STRING_LENGTH

Default maximum size of a byte string in bytes

MAX_CERTIFICATE_LENGTH

Default maximum size of a certificate to send

MAX_CHUNK_COUNT

Default maximum number of chunks in a single message. 0 is any number but this is not recommended as the default since server memory could be exhausted. Default number can be overridden by client / server config which is where it should happen if you want a different figure. In clients chunk size is a preferred value since the server can modify it during the handshake.

MAX_DECODING_DEPTH

Default maximum decoding depth for recursive data structures, i.e. if data is nested deeper than this it is an error during decoding. This is a security measure to stop deeply nested junk being sent to a server / client.

MAX_MESSAGE_SIZE

Default maximum size of a message in bytes. 0 is any length, i.e. the other end can send a message of any size which is not recommended in a server configuration. Override in the client / server config. In clients, max message size is only preferred size since it can be adjusted by the server during the handshake.

MAX_STRING_LENGTH

Default maximum size of a string in chars

SECURITY_POLICY_NONE

String used as shorthand in config files, debug etc.for None security policy

SECURITY_POLICY_NONE_URI

URI supplied for the None security policy