opcua::types::date_time

Struct DateTime

pub struct DateTime { /* private fields */ }

A date/time value. This is a wrapper around the chrono type with extra functionality for obtaining ticks in OPC UA measurements, endtimes, epoch etc.

Implementations

impl DateTime

pub fn now() -> DateTime

Constructs from the current time

pub fn now_with_offset(offset: Duration) -> DateTime

Constructs from the current time with an offset

pub fn null() -> DateTime

Creates a null date time (i.e. the epoch)

pub fn is_null(&self) -> bool

Tests if the date time is null (i.e. equal to epoch)

pub fn epoch() -> DateTime

Constructs a date time for the epoch

pub fn endtimes() -> DateTime

Constructs a date time for the endtimes

pub fn endtimes_ticks() -> i64

Returns the maximum tick value, corresponding to the end of time

pub fn ymd(year: u16, month: u16, day: u16) -> DateTime

Constructs from a year, month, day

pub fn ymd_hms( year: u16, month: u16, day: u16, hour: u16, minute: u16, second: u16, ) -> DateTime

Constructs from a year, month, day, hour, minute, second

pub fn ymd_hms_nano( year: u16, month: u16, day: u16, hour: u16, minute: u16, second: u16, nanos: u32, ) -> DateTime

Constructs from a year, month, day, hour, minute, second, nanosecond

pub fn to_rfc3339(&self) -> String

Returns an RFC 3339 and ISO 8601 date and time string such as 1996-12-19T16:39:57-08:00.

pub fn parse_from_rfc3339(s: &str) -> Result<DateTime, ()>

Parses an RFC 3339 and ISO 8601 date and time string such as 1996-12-19T16:39:57-08:00, then returns a new DateTime

pub fn ticks(&self) -> i64

Returns the time in ticks, of 100 nanosecond intervals

pub fn checked_ticks(&self) -> i64

To checked ticks. Function returns 0 or MAX_INT64 if date exceeds valid OPC UA range

pub fn as_chrono(&self) -> DateTimeUtc

Time as chrono