opcua::types

Module encoding

Contains the BinaryEncoder trait and helpers for reading and writing of scalar values and other primitives.

Structs

DecodingOptions

DepthGauge

Depth gauge is used on potentially recursive structures like Variant & ExtensionObject during decoding to limit the depth the decoder will go before giving up.

DepthLock

Depth lock holds a reference on the depth gauge. The drop ensures impl that the reference is decremented even if there is a panic unwind.

Traits

BinaryEncoder

OPC UA Binary Encoding interface. Anything that encodes to binary must implement this. It provides functions to calculate the size in bytes of the struct (for allocating memory), encoding to a stream and decoding from a stream.

Functions

byte_len_array

Calculates the length in bytes of an array of encoded type

process_decode_io_result

Converts an IO encoding error (and logs when in error) into an EncodingResult

process_encode_io_result

Converts an IO encoding error (and logs when in error) into an EncodingResult

read_array

Reads an array of the encoded type from a stream, preserving distinction between null array and empty array

read_bytes

Reads an array of bytes from the stream

read_f32

Read a 32-bit precision value from the stream

read_f64

Read a 64-bit precision from the stream

read_i16

Read an signed 16-bit value from the stream

read_i32

Read a signed 32-bit value from the stream

read_i64

Read a signed 64-bit value from the stream

read_u8

Read an unsigned byte from the stream

read_u16

Read an unsigned 16-bit value from the stream

read_u32

Read an unsigned 32-bit value from the stream

read_u64

Read an unsigned 64-bit value from the stream

write_array

Write an array of the encoded type to stream, preserving distinction between null array and empty array

write_bytes

Writes a series of identical bytes to the stream

write_f32

Writes a 32-bit precision value to the stream

write_f64

Writes a 64-bit precision value to the stream

write_i16

Writes a signed 16-bit value to the stream

write_i32

Writes a signed 32-bit value to the stream

write_i64

Writes a signed 64-bit value to the stream

write_u8

Writes an unsigned byte to the stream

write_u16

Writes an unsigned 16-bit value to the stream

write_u32

Writes an unsigned 32-bit value to the stream

write_u64

Writes an unsigned 64-bit value to the stream

Type Aliases

EncodingResult