opcua::types::encoding

Trait BinaryEncoder

pub trait BinaryEncoder<T> {
    // Required methods
    fn byte_len(&self) -> usize;
    fn encode<S: Write>(&self, stream: &mut S) -> EncodingResult<usize>;
    fn decode<S: Read>(
        stream: &mut S,
        decoding_options: &DecodingOptions,
    ) -> EncodingResult<T>;

    // Provided method
    fn encode_to_vec(&self) -> Vec<u8>  { ... }
}

OPC UA Binary Encoding interface. Anything that encodes to binary must implement this. It provides functions to calculate the size in bytes of the struct (for allocating memory), encoding to a stream and decoding from a stream.

Required Methods

fn byte_len(&self) -> usize

Returns the exact byte length of the structure as it would be if encode were called. This may be called prior to writing to ensure the correct amount of space is available.

fn encode<S: Write>(&self, stream: &mut S) -> EncodingResult<usize>

Encodes the instance to the write stream.

fn decode<S: Read>( stream: &mut S, decoding_options: &DecodingOptions, ) -> EncodingResult<T>

Decodes an instance from the read stream. The decoding options contains restrictions set by the server / client on the length of strings, arrays etc. If these limits are exceeded the implementation should return with a BadDecodingError as soon as possible.

Provided Methods

fn encode_to_vec(&self) -> Vec<u8>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.