opcua::types::node_id

Struct NodeId

pub struct NodeId {
    pub namespace: u16,
    pub identifier: Identifier,
}

An identifier for a node in the address space of an OPC UA Server.

Fields
namespace: u16

The index for a namespace

identifier: Identifier

The identifier for the node in the address space

Implementations

impl NodeId

pub fn new<T>(namespace: u16, value: T) -> NodeId
where T: 'static + Into<Identifier>,

pub fn root_folder_id() -> NodeId

Returns the node id for the root folder.

pub fn objects_folder_id() -> NodeId

Returns the node id for the objects folder.

pub fn types_folder_id() -> NodeId

Returns the node id for the types folder.

pub fn views_folder_id() -> NodeId

Returns the node id for the views folder.

pub fn is_null(&self) -> bool

Test if the node id is null, i.e. 0 namespace and 0 identifier

pub fn null() -> NodeId

Returns a null node id

pub fn next_numeric(namespace: u16) -> NodeId

pub fn as_object_id(&self) -> Result<ObjectId, NodeIdError>

Extracts an ObjectId from a node id, providing the node id holds an object id

pub fn as_reference_type_id(&self) -> Result<ReferenceTypeId, NodeIdError>

pub fn is_numeric(&self) -> bool

Test if the node id is numeric

pub fn is_string(&self) -> bool

Test if the node id is a string

pub fn is_guid(&self) -> bool

Test if the node id is a guid

pub fn is_byte_string(&self) -> bool

Test if the node id us a byte string