opcua::types::operand

Struct ContentFilterBuilder

pub struct ContentFilterBuilder { /* private fields */ }

This is a convenience for building ContentFilter using operands as building blocks This builder does not check to see that the content filter is valid, i.e. if you reference an element by index that doesn’t exist, or introduce a loop then you will not get an error until you feed it to a server and the server rejects it or breaks.

The builder takes generic types to make it easier to work with. Operands are converted to extension objects.

Implementations

impl ContentFilterBuilder

pub fn new() -> Self

pub fn eq<T, S>(self, o1: T, o2: S) -> Self
where T: Into<Operand>, S: Into<Operand>,

pub fn null<T>(self, o1: T) -> Self
where T: Into<Operand>,

pub fn gt<T, S>(self, o1: T, o2: S) -> Self
where T: Into<Operand>, S: Into<Operand>,

pub fn lt<T, S>(self, o1: T, o2: S) -> Self
where T: Into<Operand>, S: Into<Operand>,

pub fn gte<T, S>(self, o1: T, o2: S) -> Self
where T: Into<Operand>, S: Into<Operand>,

pub fn lte<T, S>(self, o1: T, o2: S) -> Self
where T: Into<Operand>, S: Into<Operand>,

pub fn like<T, S>(self, o1: T, o2: S) -> Self
where T: Into<Operand>, S: Into<Operand>,

pub fn not<T>(self, o1: T) -> Self
where T: Into<Operand>,

pub fn between<T, S, U>(self, o1: T, o2: S, o3: U) -> Self
where T: Into<Operand>, S: Into<Operand>, U: Into<Operand>,

pub fn in_list<T, S>(self, o1: T, list_items: Vec<S>) -> Self
where T: Into<Operand>, S: Into<Operand>,

pub fn and<T, S>(self, o1: T, o2: S) -> Self
where T: Into<Operand>, S: Into<Operand>,

pub fn or<T, S>(self, o1: T, o2: S) -> Self
where T: Into<Operand>, S: Into<Operand>,

pub fn cast<T, S>(self, o1: T, o2: S) -> Self
where T: Into<Operand>, S: Into<Operand>,

pub fn bitwise_and<T, S>(self, o1: T, o2: S) -> Self
where T: Into<Operand>, S: Into<Operand>,

pub fn bitwise_or<T, S>(self, o1: T, o2: S) -> Self
where T: Into<Operand>, S: Into<Operand>,

pub fn build(self) -> ContentFilter