opcua::types::qualified_name

Struct QualifiedName

pub struct QualifiedName {
    pub namespace_index: u16,
    pub name: UAString,
}

An identifier for a error or condition that is associated with a value or an operation.

A name qualified by a namespace.

For JSON, the namespace_index is saved as โUriโ and MUST be a numeric value or it will not parse. This is is in accordance with OPC UA spec that says to save the index as a numeric according to rules cut and pasted from spec below:

Name The Name component of the QualifiedName.

Uri The NamespaceIndexcomponent of the QualifiedNameencoded as a JSON number. The Urifield is omitted if the NamespaceIndex equals 0. For the non-reversible form, the NamespaceUriassociated with the NamespaceIndexportion of the QualifiedNameis encoded as JSON string unless the NamespaceIndexis 1 or if NamespaceUriis unknown. In these cases, the NamespaceIndexis encoded as a JSON number.

Fields
namespace_index: u16

The namespace index

name: UAString

The name.

Implementations

impl QualifiedName

pub fn new<T>(namespace_index: u16, name: T) -> QualifiedName
where T: Into<UAString>,

pub fn null() -> QualifiedName

pub fn is_null(&self) -> bool