opcua::types::variant

Enum Variant

pub enum Variant {

    Empty,
    Boolean(bool),
    SByte(i8),
    Byte(u8),
    Int16(i16),
    UInt16(u16),
    Int32(i32),
    UInt32(u32),
    Int64(i64),
    UInt64(u64),
    Float(f32),
    Double(f64),
    String(UAString),
    DateTime(Box<DateTime>),
    Guid(Box<Guid>),
    StatusCode(StatusCode),
    ByteString(ByteString),
    XmlElement(XmlElement),
    QualifiedName(Box<QualifiedName>),
    LocalizedText(Box<LocalizedText>),
    NodeId(Box<NodeId>),
    ExpandedNodeId(Box<ExpandedNodeId>),
    ExtensionObject(Box<ExtensionObject>),
    Variant(Box<Variant>),
    DataValue(Box<DataValue>),
    DiagnosticInfo(Box<DiagnosticInfo>),
    Array(Box<Array>),

}

A Variant holds built-in OPC UA data types, including single and multi dimensional arrays, data values and extension objects.

As variants may be passed around a lot on the stack, Boxes are used for more complex types to keep the size of this type down a bit, especially when used in arrays.

Variants

Empty

Empty type has no value. It is equivalent to a Null value (part 6 5.1.6)

Boolean(bool)

Boolean

SByte(i8)

Signed byte

Byte(u8)

Unsigned byte

Int16(i16)

Signed 16-bit int

UInt16(u16)

Unsigned 16-bit int

Int32(i32)

Signed 32-bit int

UInt32(u32)

Unsigned 32-bit int

Int64(i64)

Signed 64-bit int

UInt64(u64)

Unsigned 64-bit int

Float(f32)

Float

Double(f64)

Double

String(UAString)

String

DateTime(Box<DateTime>)

DateTime

Guid(Box<Guid>)

Guid

StatusCode(StatusCode)

StatusCode

ByteString(ByteString)

ByteString

XmlElement(XmlElement)

XmlElement

QualifiedName(Box<QualifiedName>)

QualifiedName

LocalizedText(Box<LocalizedText>)

LocalizedText

NodeId(Box<NodeId>)

NodeId

ExpandedNodeId(Box<ExpandedNodeId>)

ExpandedNodeId

ExtensionObject(Box<ExtensionObject>)

ExtensionObject

Variant(Box<Variant>)

DataValue(Box<DataValue>)

DiagnosticInfo(Box<DiagnosticInfo>)

Array(Box<Array>)

Single dimension array which can contain any scalar type, all the same type. Nested arrays will be rejected.

Implementations

impl Variant

pub fn test_encoding_flag(encoding_mask: u8, flag: u8) -> bool

Test the flag (convenience method)

pub fn cast(&self, target_type: VariantTypeId) -> Variant

Performs an EXPLICIT cast from one type to another. This will first attempt an implicit conversion and only then attempt to cast. Casting is potentially lossy.

pub fn convert(&self, target_type: VariantTypeId) -> Variant

Performs an IMPLICIT conversion from one type to another

pub fn type_id(&self) -> VariantTypeId

pub fn is_numeric(&self) -> bool

Tests and returns true if the variant holds a numeric type

pub fn is_array(&self) -> bool

Test if the variant holds an array

pub fn is_array_of_type(&self, variant_type: VariantTypeId) -> bool

pub fn is_valid(&self) -> bool

Tests that the variant is in a valid state. In particular for arrays ensuring that the values are all acceptable and for a multi dimensional array that the dimensions equal the actual values.

pub fn as_f64(&self) -> Option<f64>

Converts the numeric type to a double or returns None

pub fn array_data_type(&self) -> Option<NodeId>

pub fn scalar_data_type(&self) -> Option<NodeId>

pub fn to_byte_array(&self) -> Result<Self, StatusCode>

This function is for a special edge case of converting a byte string to a single array of bytes

pub fn eq_scalar_type(&self, other: &Variant) -> bool

pub fn eq_array_type(&self, other: &Variant) -> bool

pub fn set_range_of( &mut self, range: NumericRange, other: &Variant, ) -> Result<(), StatusCode>

pub fn range_of(&self, range: NumericRange) -> Result<Variant, StatusCode>

This function gets a range of values from the variant if it is an array, or returns a clone of the variant itself.