opcua::types::status_code

Struct StatusCode

pub struct StatusCode(/* private fields */);

Implementations

impl StatusCode

pub const STATUS_MASK: Self

pub const BIT_MASK: Self

pub const IS_ERROR: Self

pub const IS_UNCERTAIN: Self

pub const HISTORICAL_RAW: Self

pub const HISTORICAL_CALCULATED: Self

pub const HISTORICAL_INTERPOLATED: Self

pub const HISTORICAL_RESERVED: Self

pub const HISTORICAL_PARTIAL: Self

pub const HISTORICAL_EXTRA_DATA: Self

pub const HISTORICAL_MULTI_VALUE: Self

pub const OVERFLOW: Self

pub const LIMIT_LOW: Self

pub const LIMIT_HIGH: Self

pub const LIMIT_CONSTANT: Self

pub const LIMIT_DATA_VALUE: Self

pub const SEMANTICS_CHANGED: Self

pub const STRUCTURE_CHANGED: Self

pub const Good: Self

pub const GoodSubscriptionTransferred: Self

pub const GoodCompletesAsynchronously: Self

pub const GoodOverload: Self

pub const GoodClamped: Self

pub const GoodLocalOverride: Self

pub const GoodEntryInserted: Self

pub const GoodEntryReplaced: Self

pub const GoodNoData: Self

pub const GoodMoreData: Self

pub const GoodCommunicationEvent: Self

pub const GoodShutdownEvent: Self

pub const GoodCallAgain: Self

pub const GoodNonCriticalTimeout: Self

pub const GoodResultsMayBeIncomplete: Self

pub const GoodDataIgnored: Self

pub const GoodEdited: Self

pub const GoodPostActionFailed: Self

pub const GoodDependentValueChanged: Self

pub const GoodEdited_DependentValueChanged: Self

pub const GoodEdited_DominantValueChanged: Self

pub const GoodEdited_DominantValueChanged_DependentValueChanged: Self

pub const UncertainReferenceOutOfServer: Self

pub const UncertainNoCommunicationLastUsableValue: Self

pub const UncertainLastUsableValue: Self

pub const UncertainSubstituteValue: Self

pub const UncertainInitialValue: Self

pub const UncertainSensorNotAccurate: Self

pub const UncertainEngineeringUnitsExceeded: Self

pub const UncertainSubNormal: Self

pub const UncertainDataSubNormal: Self

pub const UncertainReferenceNotDeleted: Self

pub const UncertainNotAllNodesAvailable: Self

pub const UncertainDominantValueChanged: Self

pub const UncertainDependentValueChanged: Self

pub const BadUnexpectedError: Self

pub const BadInternalError: Self

pub const BadOutOfMemory: Self

pub const BadResourceUnavailable: Self

pub const BadCommunicationError: Self

pub const BadEncodingError: Self

pub const BadDecodingError: Self

pub const BadEncodingLimitsExceeded: Self

pub const BadUnknownResponse: Self

pub const BadTimeout: Self

pub const BadServiceUnsupported: Self

pub const BadShutdown: Self

pub const BadServerNotConnected: Self

pub const BadServerHalted: Self

pub const BadNothingToDo: Self

pub const BadTooManyOperations: Self

pub const BadDataTypeIdUnknown: Self

pub const BadCertificateInvalid: Self

pub const BadSecurityChecksFailed: Self

pub const BadCertificateTimeInvalid: Self

pub const BadCertificateIssuerTimeInvalid: Self

pub const BadCertificateHostNameInvalid: Self

pub const BadCertificateUriInvalid: Self

pub const BadCertificateUseNotAllowed: Self

pub const BadCertificateIssuerUseNotAllowed: Self

pub const BadCertificateUntrusted: Self

pub const BadCertificateRevocationUnknown: Self

pub const BadCertificateIssuerRevocationUnknown: Self

pub const BadCertificateRevoked: Self

pub const BadCertificateIssuerRevoked: Self

pub const BadUserAccessDenied: Self

pub const BadIdentityTokenInvalid: Self

pub const BadIdentityTokenRejected: Self

pub const BadSecureChannelIdInvalid: Self

pub const BadInvalidTimestamp: Self

pub const BadNonceInvalid: Self

pub const BadSessionIdInvalid: Self

pub const BadSessionClosed: Self

pub const BadSessionNotActivated: Self

pub const BadSubscriptionIdInvalid: Self

pub const BadRequestHeaderInvalid: Self

pub const BadTimestampsToReturnInvalid: Self

pub const BadRequestCancelledByClient: Self

pub const BadNoCommunication: Self

pub const BadWaitingForInitialData: Self

pub const BadNodeIdInvalid: Self

pub const BadNodeIdUnknown: Self

pub const BadAttributeIdInvalid: Self

pub const BadIndexRangeInvalid: Self

pub const BadIndexRangeNoData: Self

pub const BadDataEncodingInvalid: Self

pub const BadDataEncodingUnsupported: Self

pub const BadNotReadable: Self

pub const BadNotWritable: Self

pub const BadOutOfRange: Self

pub const BadNotSupported: Self

pub const BadNotFound: Self

pub const BadObjectDeleted: Self

pub const BadNotImplemented: Self

pub const BadMonitoringModeInvalid: Self

pub const BadMonitoredItemIdInvalid: Self

pub const BadMonitoredItemFilterInvalid: Self

pub const BadMonitoredItemFilterUnsupported: Self

pub const BadFilterNotAllowed: Self

pub const BadStructureMissing: Self

pub const BadEventFilterInvalid: Self

pub const BadContentFilterInvalid: Self

pub const BadFilterOperandInvalid: Self

pub const BadContinuationPointInvalid: Self

pub const BadNoContinuationPoints: Self

pub const BadReferenceTypeIdInvalid: Self

pub const BadBrowseDirectionInvalid: Self

pub const BadNodeNotInView: Self

pub const BadServerUriInvalid: Self

pub const BadServerNameMissing: Self

pub const BadDiscoveryUrlMissing: Self

pub const BadSempahoreFileMissing: Self

pub const BadRequestTypeInvalid: Self

pub const BadSecurityModeRejected: Self

pub const BadSecurityPolicyRejected: Self

pub const BadTooManySessions: Self

pub const BadUserSignatureInvalid: Self

pub const BadApplicationSignatureInvalid: Self

pub const BadNoValidCertificates: Self

pub const BadRequestCancelledByRequest: Self

pub const BadParentNodeIdInvalid: Self

pub const BadReferenceNotAllowed: Self

pub const BadNodeIdRejected: Self

pub const BadNodeIdExists: Self

pub const BadNodeClassInvalid: Self

pub const BadBrowseNameInvalid: Self

pub const BadBrowseNameDuplicated: Self

pub const BadNodeAttributesInvalid: Self

pub const BadTypeDefinitionInvalid: Self

pub const BadSourceNodeIdInvalid: Self

pub const BadTargetNodeIdInvalid: Self

pub const BadDuplicateReferenceNotAllowed: Self

pub const BadInvalidSelfReference: Self

pub const BadReferenceLocalOnly: Self

pub const BadNoDeleteRights: Self

pub const BadServerIndexInvalid: Self

pub const BadViewIdUnknown: Self

pub const BadTooManyMatches: Self

pub const BadQueryTooComplex: Self

pub const BadNoMatch: Self

pub const BadMaxAgeInvalid: Self

pub const BadHistoryOperationInvalid: Self

pub const BadHistoryOperationUnsupported: Self

pub const BadWriteNotSupported: Self

pub const BadTypeMismatch: Self

pub const BadMethodInvalid: Self

pub const BadArgumentsMissing: Self

pub const BadTooManySubscriptions: Self

pub const BadTooManyPublishRequests: Self

pub const BadNoSubscription: Self

pub const BadSequenceNumberUnknown: Self

pub const BadMessageNotAvailable: Self

pub const BadInsufficientClientProfile: Self

pub const BadTcpServerTooBusy: Self

pub const BadTcpMessageTypeInvalid: Self

pub const BadTcpSecureChannelUnknown: Self

pub const BadTcpMessageTooLarge: Self

pub const BadTcpNotEnoughResources: Self

pub const BadTcpInternalError: Self

pub const BadTcpEndpointUrlInvalid: Self

pub const BadRequestInterrupted: Self

pub const BadRequestTimeout: Self

pub const BadSecureChannelClosed: Self

pub const BadSecureChannelTokenUnknown: Self

pub const BadSequenceNumberInvalid: Self

pub const BadConfigurationError: Self

pub const BadNotConnected: Self

pub const BadDeviceFailure: Self

pub const BadSensorFailure: Self

pub const BadOutOfService: Self

pub const BadDeadbandFilterInvalid: Self

pub const BadRefreshInProgress: Self

pub const BadConditionAlreadyDisabled: Self

pub const BadConditionDisabled: Self

pub const BadEventIdUnknown: Self

pub const BadNoData: Self

pub const BadDataLost: Self

pub const BadDataUnavailable: Self

pub const BadEntryExists: Self

pub const BadNoEntryExists: Self

pub const BadTimestampNotSupported: Self

pub const BadInvalidArgument: Self

pub const BadConnectionRejected: Self

pub const BadDisconnect: Self

pub const BadConnectionClosed: Self

pub const BadInvalidState: Self

pub const BadEndOfStream: Self

pub const BadNoDataAvailable: Self

pub const BadWaitingForResponse: Self

pub const BadOperationAbandoned: Self

pub const BadExpectedStreamToBlock: Self

pub const BadWouldBlock: Self

pub const BadSyntaxError: Self

pub const BadMaxConnectionsReached: Self

pub const BadRequestTooLarge: Self

pub const BadResponseTooLarge: Self

pub const BadEventNotAcknowledgeable: Self

pub const BadInvalidTimestampArgument: Self

pub const BadProtocolVersionUnsupported: Self

pub const BadStateNotActive: Self

pub const BadFilterOperatorInvalid: Self

pub const BadFilterOperatorUnsupported: Self

pub const BadFilterOperandCountMismatch: Self

pub const BadFilterElementInvalid: Self

pub const BadFilterLiteralInvalid: Self

pub const BadIdentityChangeNotSupported: Self

pub const BadNotTypeDefinition: Self

pub const BadViewTimestampInvalid: Self

pub const BadViewParameterMismatch: Self

pub const BadViewVersionInvalid: Self

pub const BadConditionAlreadyEnabled: Self

pub const BadDialogNotActive: Self

pub const BadDialogResponseInvalid: Self

pub const BadConditionBranchAlreadyAcked: Self

pub const BadConditionBranchAlreadyConfirmed: Self

pub const BadConditionAlreadyShelved: Self

pub const BadConditionNotShelved: Self

pub const BadShelvingTimeOutOfRange: Self

pub const BadAggregateListMismatch: Self

pub const BadAggregateNotSupported: Self

pub const BadAggregateInvalidInputs: Self

pub const BadBoundNotFound: Self

pub const BadBoundNotSupported: Self

pub const BadAggregateConfigurationRejected: Self

pub const BadTooManyMonitoredItems: Self

pub const BadDominantValueChanged: Self

pub const BadDependentValueChanged: Self

pub const BadRequestNotAllowed: Self

pub const BadTooManyArguments: Self

pub const BadSecurityModeInsufficient: Self

pub const BadCertificateChainIncomplete: Self

pub const BadLicenseExpired: Self

pub const BadLicenseLimitsExceeded: Self

pub const BadLicenseNotAvailable: Self

pub const BadNotExecutable: Self

pub const BadNumericOverflow: Self

pub const BadRequestNotComplete: Self

pub const BadCertificatePolicyCheckFailed: Self

pub const BadAlreadyExists: Self

pub const BadEdited_OutOfRange: Self

pub const BadInitialValue_OutOfRange: Self

pub const BadOutOfRange_DominantValueChanged: Self

pub const BadEdited_OutOfRange_DominantValueChanged: Self

pub const BadOutOfRange_DominantValueChanged_DependentValueChanged: Self

pub const BadEdited_OutOfRange_DominantValueChanged_DependentValueChanged: Self

impl StatusCode

pub const fn empty() -> Self

Get a flags value with all bits unset.

pub const fn all() -> Self

Get a flags value with all known bits set.

pub const fn bits(&self) -> u32

Get the underlying bits value.

The returned value is exactly the bits set in this flags value.

pub const fn from_bits(bits: u32) -> Option<Self>

Convert from a bits value.

This method will return None if any unknown bits are set.

pub const fn from_bits_truncate(bits: u32) -> Self

Convert from a bits value, unsetting any unknown bits.

pub const fn from_bits_retain(bits: u32) -> Self

Convert from a bits value exactly.

pub fn from_name(name: &str) -> Option<Self>

Get a flags value with the bits of a flag with the given name set.

This method will return None if name is empty or doesnโt correspond to any named flag.

pub const fn is_empty(&self) -> bool

Whether all bits in this flags value are unset.

pub const fn is_all(&self) -> bool

Whether all known bits in this flags value are set.

pub const fn intersects(&self, other: Self) -> bool

Whether any set bits in a source flags value are also set in a target flags value.

pub const fn contains(&self, other: Self) -> bool

Whether all set bits in a source flags value are also set in a target flags value.

pub fn insert(&mut self, other: Self)

The bitwise or (|) of the bits in two flags values.

pub fn remove(&mut self, other: Self)

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. remove wonโt truncate other, but the ! operator will.

pub fn toggle(&mut self, other: Self)

The bitwise exclusive-or (^) of the bits in two flags values.

pub fn set(&mut self, other: Self, value: bool)

Call insert when value is true or remove when value is false.

pub const fn intersection(self, other: Self) -> Self

The bitwise and (&) of the bits in two flags values.

pub const fn union(self, other: Self) -> Self

The bitwise or (|) of the bits in two flags values.

pub const fn difference(self, other: Self) -> Self

The intersection of a source flags value with the complement of a target flags value (&!).

This method is not equivalent to self & !other when other has unknown bits set. difference wonโt truncate other, but the ! operator will.

pub const fn symmetric_difference(self, other: Self) -> Self

The bitwise exclusive-or (^) of the bits in two flags values.

pub const fn complement(self) -> Self

The bitwise negation (!) of the bits in a flags value, truncating the result.

impl StatusCode

pub const fn iter(&self) -> Iter<StatusCode>

Yield a set of contained flags values.

Each yielded flags value will correspond to a defined named flag. Any unknown bits will be yielded together as a final flags value.

pub const fn iter_names(&self) -> IterNames<StatusCode>

Yield a set of contained named flags values.

This method is like iter, except only yields bits in contained named flags. Any unknown bits, or bits not corresponding to a contained flag will not be yielded.

impl StatusCode

pub fn name(self) -> &'static str

Returns the descriptive name for the status code, e.g. to put a meaningful code in a log file

pub fn description(self) -> &'static str

Returns the descriptive text for the status code

pub fn from_u32(code: u32) -> Option<StatusCode>

Takes an OPC UA status code as a UInt32 and returns the matching StatusCode, assuming there is one Note that this is lossy since any bits associated with the status code will be ignored.

impl StatusCode

pub fn bitflags(&self) -> StatusCode

Returns the bit flags of the status code, i.e. it masks out the actual status code value

pub fn status(&self) -> StatusCode

Returns the status only, i.e. it masks out any bit flags that come with the status code

pub fn is_bad(&self) -> bool

Tests if the status code is bad

pub fn is_uncertain(&self) -> bool

Tests if the status code is uncertain

pub fn is_good(&self) -> bool

Tests if the status code is good (i.e. not bad or uncertain)